Add files via upload
The bus booking system is a Java Spring Boot, MySQL and Angular Project which runs on the tomcat server, you can also run this project eclipse and Spring Tool Suites. We have developed this java Spring Boot and Angular project on Bus Booking System to automate the Bus Booking System process. The main features of this project are to manage Bus, Ticket Booking, Sales, Operations, Customers and Booking. 
 The Online Bus Booking System is a web-based application that allows visitors to check ticket availability, buy a bus ticket and pay for the bus ticket online.